requirements.txt
aerich==0.7.1
aiomysql==0.1.1
aioredis==2.0.1
fastapi==0.88.0
passlib==1.7.4
bcrypt==4.0.1
tortoise-orm==0.19.2
uvicorn==0.20.0
